+++ src/components/BulkEditPane/BulkEditListResult/BulkEditDeleteModal/index.js
@@ -1,0 +1,2 @@
+export { default as BulkEditDeleteModal } from './BulkEditDeleteModal';
+