+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * [UIBULKED-766](https://folio-org.atlassian.net/browse/UIBULKED-766) Bulk Edit - Support editing material type in item records.
 * [UIBULKED-590](https://folio-org.atlassian.net/browse/UIBULKED-590) Replace "Remove all"  operation on MARC instance edits with "Remove field" and "Remove subfield".
 * [UIBULKED-99](https://folio-org.atlassian.net/browse/UIBULKED-99) User records bulk delete authorization.
+* [UIBULKED-26](https://folio-org.atlassian.net/browse/UIBULKED-26) Bulk delete user records - Delete user records? form.
+* [UIBULKED-27](https://folio-org.atlassian.net/browse/UIBULKED-27) Confirmation form - Error accordion.
 
 ## [5.1.1](https://github.com/folio-org/ui-bulk-edit/tree/v5.1.1) (2026-05-14)